Fixing problems with units

+++ src/toolbox.cpp
@@ -50,7 +50,8 @@ int findLastStep(const char *path) {
 /* 
  * Identify the snowpack mesh based on file 0.txt 
  */
-void checkmesh2(int* H_local,int* L_local,int* h_layer,int* l_layer,int* nh,int* nl,std::ofstream* logPULSEfile)
+void checkmesh2(double* H_local,double* L_local,double* h_layer,double* l_layer,
+                int* nh,int* nl,std::ofstream* logPULSEfile)
 {
     unsigned int a, timstart;
