+++ src/toolbox.h
@@ -14,6 +14,7 @@ std::string SplitFilename (const std::string& str);
 
 int findLastStep(const char *path);
 
-void checkmesh2(int* H_local,int* L_local,int* h_layer,int* l_layer,int* nh,int* nl,std::ofstream* logPULSEfile);
+void checkmesh2(double* H_local,double* L_local,double* h_layer,double* l_layer,
+                int* nh,int* nl,std::ofstream* logPULSEfile);
 
 #endif
